Add Cai's example of prettyprinter breakage
That's not how Cai originally presented the example, but I think that
prettyprinting should work also for such examples.
Hence, as Tillmann originally suspected, we need to rename variables.
In the future, I plan to fix this. Since the goal is to make ASTs
more readable, I plan not to rename every name, and instead to keep
names unchanged when it's not a problem. "Problem" is ambiguous on
purpose there: I think that every shadowing should be removed, but I'm
not entirely sure.