Move lambda-calculus implementation into client
Instead of having it in the project defining the macro, it should only
be in the project using macros. This reduces tremendously the overhead
of setting up a separate project for macros - in fact, you simply need
to import the library from the macros project, mix the macro component
with the calculus component, and use its macro members.

Next, I will refine the interface that a calculus component must supply.